refact(composables): create useCollaborationInvite and migrate InviteUserDialog fetchWithAuth (#6091)
